cloud-function/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery==2.3.1
pytz